+++ TablePro/Views/Sidebar/DatabaseTreeSelection.swift
@@ -17,7 +17,7 @@ internal enum DatabaseTreeSelection {
         switch kind {
         case .status, .recentSection, .objectKindSection, .hierarchicalSchemaSection, .redisKeysSection:
             return false
-        case .database, .schema, .table, .routine, .recentTable, .redisNode:
+        case .database, .schema, .table, .routine, .recentTable, .containerObjectKindSection, .redisNode:
             return true
         }
     }